+++ tests/unit/components/pokemon/PokemonItem.test.js
@@ -7,11 +7,6 @@ jest.mock('@/components/ui/BaseButton.vue', () => ({
   template: '<div class="mocked-base-button"></div>',
 }));
 
-jest.mock('@/components/ui/CenteredColumn.vue', () => ({
-  name: 'CenteredColumn',
-  template: '<div class="mocked-centered-column"></div>',
-}));
-
 jest.mock('@/components/pokemon/PokemonItemHeader.vue', () => ({
   name: 'PokemonItemHeader',
   template: '<div class="mocked-pokemon-item-header"></div>',
@@ -84,7 +79,6 @@ describe('PokemonItem', () => {
 
   it('renders its respective components', () => {
     expect(wrapper.find('basebutton-stub').exists()).toBe(true);
-    expect(wrapper.find('centeredcolumn-stub').exists()).toBe(true);
     expect(wrapper.find('pokemonitemheader-stub').exists()).toBe(false);
     expect(wrapper.find('pokemonitemstats-stub').exists()).toBe(false);
     expect(wrapper.find('pokemonitemtype-stub').exists()).toBe(false);